add form load time vs submission time check
adding a hidden field to the contact form did not seem to stop the spam emails.

So, I decided to add a submission time check. If the submit button is clicked less than 5 seconds after the page is loaded, it will not submit the form.

It was my belief these bots auto populate text, and click submit instantly.

Tested and validated, and it's been 5 days since the prod change, and not a single spam email (usually 2 per day)